requirements.txt
gunicorn==20.0.4
Flask==1.0.0
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
simplejson==3.16.0
google-cloud-storage==1.16.1
python-resize-image==1.1.19
firebase-admin==2.17.0
google-api-python-client==1.7.9
geopy==1.22.0
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.3.1
idna==2.7
Jinja2==2.10
PyMySQL==0.9.0
six==1.11.0
Werkzeug==0.14.1